refactor: Refactor Control flow classes and improve Documentation
