+++ src/astx/flows.py
@@ -1,3 +1,4 @@
+"""Module for controle flow AST."""
 from typing import List, Tuple, Optional
 
 from astx.base import StatementType, Expr, SourceLocation, ASTKind
@@ -52,25 +53,3 @@ class ForStmt(StatementType):
         self.step = step
         self.body = body
         self.kind = ASTKind.ForKind
-
-
-class VarExprAST(StatementType):
-    """AST class for variable declaration."""
-
-    var_names: List[Tuple[str, Expr]]
-    type_name: str
-    body: Block
-
-    def __init__(
-        self,
-        var_names: List[Tuple[str, Expr]],
-        type_name: str,
-        body: Block,
-        loc: SourceLocation = SourceLocation(0, 0),
-    ) -> None:
-        """Initialize the VarExprAST instance."""
-        self.loc = loc
-        self.var_names = var_names
-        self.type_name = type_name
-        self.body = body
-        self.kind = ASTKind.VarKind
